% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook-package.R
\docType{package}
\name{codebook-package}
\alias{codebook-package}
\title{Create an interactive codebook}
\source{
Rho's web-codebook JavaScript library: \url{https://github.com/RhoInc/web-codebook}.
}
\description{
The \code{codebook} package provides a simple interactive framework for exploring data.  
The package serves as an R interface for Rho's \code{web-codebook} 
JavaScript library. This tool provides a concise summary of every variable in a data frame and 
includes interactive features such as real-time filters, grouping, and highlighting. 
The R interface allows the analyst to interactively explore datasets within the typical 
analysis workflow and working environment.  
\code{codebook} has been built using the [htmlwidgets](htmlwidgets.org) framework.
}
