% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebookApp.R
\name{codebookApp}
\alias{codebookApp}
\title{Codebook Shiny App and RStudio Add-in}
\usage{
codebookApp()
}
\description{
The codebook shiny application produces an interactive codebook using data from your R environment 
or a file upload. Optionally select to view a static summary of the data using Hmisc::describe(). 
The resulting codebook (interactive or static) may be downloaded as an HTML file.
}
