% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorerApp.R
\name{explorerApp}
\alias{explorerApp}
\title{Codebook Explorer Shiny App and RStudio Add-in}
\usage{
explorerApp()
}
\description{
The explorer shiny application generates an interactive codebook explorer 
using all datasets currently loaded in the user's R session. If no data are available, 
the app is populated using the data from the datasets package. The user can upload one or 
more sas7bdat or csv files by navigating to a directory and selecting individual file(s).
}
