% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Create an interactive codebook}
\usage{
codebook(data, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A data frame containing the Adverse Events data.}
}
\description{
This function creates an interactive codebook using R htmlwidgets.
}
\examples{
webcodebook(data=mtcars)

}
