% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer.R
\name{explorer}
\alias{explorer}
\title{Create an interactive codebook}
\usage{
explorer(dataList = NULL, addEnv = TRUE, demo = FALSE)
}
\arguments{
\item{dataList}{A list of (optionally named) data frames. Also accepts a character vector of data frame names (must be loaded in environment).}

\item{addEnv}{Logical. Indicates whether to add all data frames in current environemnt to explorer. Defaults to \code{addEnv=TRUE}.}

\item{demo}{Logical. Indicates whether to display demo data frames.  If \code{TRUE}, \code{dataList} and \code{addEnv} settings will be ignored.  Defaults to \code{demo=FALSE}.}
}
\description{
This function creates a series of interactive codebook using R htmlwidgets.
}
\examples{
explorer(dataList = list(Cars = mtcars, Iris = iris))

explorer(dataList = c("mtcars", "iris"))

}
