% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer.R
\name{explorer}
\alias{explorer}
\title{Create an interactive codebook}
\usage{
explorer(dataList = c(), addEnv = T, demo = F)
}
\arguments{
\item{addEnv}{boolean. Indicates whether to add all data frames in current environemnt to explorer.}

\item{dataArray}{a list of dataframes (TODO change to tibble?).}
}
\description{
This function creates a series of interactive codebook using R htmlwidgets.
}
\examples{
explorer(dataArray=c(cars, iris))

}
